chore: moved helper function to utils
