+++ packages/shared/lib/auxiliary/deep-link/utils/getAmountFromSearchParam.ts
