+++ packages/shared/lib/auxiliary/deep-link/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './getAmountFromSearchParam'
